This is C code and neccesary files for Pass 1 of my Assembler.
The files Input.txt and Optab.txt are user defined.
Input.txt file contains a program for which we have to find the value of
locations counter in each step and to generate a Symtab.txt , Length.txt
file and an intermediate file Output.txt
In Optab.txt file i have mentioned mnemonics name,their size and the
opcode for the particular mnemonic seperately.
A file named Symtab.txt is created which has information about the
operand name and the Address assigned to them.
Length.txt contains length of Program.
Output of program is generated in the Intermediate file.